Move entityType helpers to metabase/utils
