+++ frontend/src/metabase/redux/store/entities.ts
@@ -12,19 +12,6 @@ import type {
   NormalizedTable,
 } from "metabase-types/api";
 
-// backend returns model = "card" instead of "question"
-export const entityTypeForModel = (model: string): string => {
-  if (model === "card" || model === "dataset" || model === "metric") {
-    return "questions";
-  }
-  return `${model}s`;
-};
-
-export const entityTypeForObject = (
-  object?: { model: string } | null,
-): string | undefined =>
-  object ? entityTypeForModel(object.model) : undefined;
-
 export interface EntitiesState {
   collections: Record<string, NormalizedCollection>;
   dashboards: Record<string, NormalizedDashboard>;
