+++ frontend/src/metabase/schema.js
@@ -2,7 +2,7 @@
 
 import { schema } from "normalizr";
 
-import { entityTypeForObject } from "metabase/redux/store/entities";
+import { entityTypeForObject } from "metabase/utils/entity-types";
 import { getUniqueFieldId } from "metabase-lib/v1/metadata/utils/fields";
 import { SAVED_QUESTIONS_VIRTUAL_DB_ID } from "metabase-lib/v1/metadata/utils/saved-questions";
 import { generateSchemaId } from "metabase-lib/v1/metadata/utils/schema";
